pyproject.toml
[tool]
[tool.poetry]
name = "puzzup-2023"
version = "0.0.0"
description = "Puzzle development webapp"
authors = ["betaveros", "mitchgu", "jakob223", "dvorak42", "fortenforge", "Sandy Weisz", "James Sugrono", "teammate"]

[tool.poetry.dependencies]
python = "==3.9.*"
asgiref = "==3.2.10"
bleach = "==3.1.5"
certifi = "==2020.6.20"
click = "==7.1.2"
cycler = "==0.10.0"
discord-interactions = "==0.2.0"
dj-database-url = "==0.5.0"
django = "==3.1.13"
django-debug-toolbar = "3.2.4"
django-filter = "==2.4.0"
django-import-export = "==2.7.1"
djangorestframework = "==3.12.2"
gevent = "==20.6.2"
gitdb = "==4.0.5"
gitpython = "==3.1.7"
greenlet = "==0.4.16"
gunicorn = "==20.0.4"
importlib-metadata = "==1.7.0"
invoke = "==1.6.0"
ipython = "==8.0.1"
kiwisolver = "==1.2.0"
markdown = "==3.2.2"
matplotlib = "==3.5.1"
newrelic = "==5.16.1.146"
numpy = "==1.22.1"
pip-tools = "==5.3.1"
postmarker = "==1.0"
psycopg2-binary = "==2.9.3"
pydantic = "==1.9.0"
pyparsing = "==2.4.7"
python-dateutil = "==2.8.1"
python-decouple = "==3.5"
python-dotenv = "==0.19.2"
pytz = "==2020.1"
pyyaml = "==6.0"
requests = "==2.25.1"
sentry-sdk = "==1.5.4"
six = "==1.15.0"
smmap = "==3.0.4"
sqlparse = "==0.3.1"
webencodings = "==0.5.1"
whitenoise = "==5.3.0"
zipp = "==3.1.0"
"zope.event" = "==4.4"
"zope.interface" = "==5.1.0"
beautifulsoup4 = "==4.10.0"
google-api-python-client = "==2.37.0"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
black = "==21.12b0"
pre-commit = "==2.17.0"
